requirements.txt
pydub
srt
opencc
openai-whisper
torchaudio
tqdm